% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\alias{ognd_boxplot}
\title{box_plot}
\usage{
ognd_boxplot(measure, dimension, dataframe = NULL, alpha = 0.66,
  notch = T, title = NULL, x_label = NULL, y_label = NULL,
  facet_formula = NULL)
}
\description{
box_plot
}

