% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatter plot or bubble chart}
\usage{
scatter_plot(x, y, dataframe = NULL, color = NULL, size = NULL,
  alpha = 0.66, title = NULL, x_label = NULL, y_label = NULL,
  color_title = NULL, size_title = NULL, facet_formula = NULL)
}
\arguments{
\item{x}{The x axis variable}

\item{y}{The y axis variable}

\item{dataframe}{An optional dataframe. If dataframe is used, all variables
(x, y, color and size) must be of the dataframe.}

\item{color}{A vector of the same length of x and y. An arbitrary color
(e.g. "blue") cannot be used. If vector color is not provided, the
default color is blue.}

\item{size}{A vector of the same length of x and y. An arbitrary size (e.g.
8) cannot be used.}

\item{alpha}{The transparency of the data points.}

\item{title}{The title of the plot}

\item{x_label}{The x axis label. If no label is given, the name of the x
variable is used.}

\item{y_label}{The y axis label. If no label is given, the name of the y
variable is used.}

\item{color_title}{The title of the color legend. If there is no color
vector, the color_title is ignored.}

\item{size_title}{The title of the size legend. If there is no size vector,
the size_title is ignored.}

\item{facet_formula}{The formula to create facet grids on ggplot.}
}
\description{
Wrapper for ggplot() + geom_point() visualisations
}
\details{
This function allows the creation of scatter plots and bubble
         charts in one line. It handles color and size automatically
         using predefined settings.
}
\examples{
# Load packages
library(gridExtra)

# Set a seed to always have the same results
set.seed(0010)

# Create vectors
var_x = sort(rnorm(100))
var_y = var_x * 2 + rnorm(100)
var_z = c(rep('A', 25), rep('B', 25), rep('C', 25), rep('D', 25))
var_s = rnorm(100)
var_f1 = sample(x = c('Alpha', 'Bravo'), size = 100, replace = T)
var_f2 = sample(x = c('Sierra', 'Tango'), size = 100, replace = T)

# Create dataframe
df = data.frame(x = var_x, y = var_y, z = var_z, s = var_s,
                a = var_f1, b = var_f2)

# Create scatter plots
p1 = scatter_plot(x = var_x, y = var_y,
                  title = 'Simple')
p2 = scatter_plot(x = var_x, y = var_y, color = var_z,
                  title = 'With categorical color')
p3 = scatter_plot(x = var_x, y = var_y, size = var_s,
                  title = 'With size')
p4 = scatter_plot(x = var_x, y = var_y, color = var_x, size = var_s,
                  title = 'With continuous color and size')

# Plot them
grid.arrange(p1, p2, p3, p4, ncol = 2)

# Wait for key stroke
key_stroke = readline(prompt = 'Press Enter to continue...')

# Plot facets
scatter_plot(x = x,
             y = y,
             dataframe = df,
             color = z,
             size = s,
             title = 'All plus df, labels and facets',
             x_label = 'My X Var',
             y_label = 'My Y Var',
             color_title = 'Color',
             size_title = 'Size',
             facet_formula = a ~ b)
}

